move explore controller method to itinerary
